refactor: 코드 리팩토링, 엔티티 맵 수정
