+++ build.gradle
@@ -5,7 +5,7 @@ plugins {
 }
 
 group = 'com.hanghae99'
-version = '1.0.0-SNAPSHOT-'+new Date().format("yyyyMMddHHmmss")
+version = '1.0.1-SNAPSHOT-'+new Date().format("yyyyMMddHHmmss")
 sourceCompatibility = '1.8'
 
 configurations {
