+++ src/main/java/com/hanghae99/afterwork/controller/ProfileController.java
@@ -8,8 +8,8 @@ import org.springframework.web.bind.annotation.RestController;
 import java.util.Arrays;
 import java.util.List;
 
-@RequiredArgsConstructor
 @RestController
+@RequiredArgsConstructor
 public class ProfileController {
     private final Environment env;
