+++ src/main/java/com/hanghae99/afterwork/dto/CategoryResponseDto.java
@@ -1,7 +1,7 @@
 package com.hanghae99.afterwork.dto;
 
-import com.hanghae99.afterwork.model.Interest;
-import com.hanghae99.afterwork.model.Product;
+import com.hanghae99.afterwork.entity.Interest;
+import com.hanghae99.afterwork.entity.Product;
 import lombok.Data;
 
 import java.util.ArrayList;
