+++ src/main/java/com/hanghae99/afterwork/dto/ProductResponseDto.java
@@ -1,8 +1,8 @@
 package com.hanghae99.afterwork.dto;
 
-import com.hanghae99.afterwork.model.Category;
-import com.hanghae99.afterwork.model.Collect;
-import com.hanghae99.afterwork.model.Product;
+import com.hanghae99.afterwork.entity.Category;
+import com.hanghae99.afterwork.entity.Collect;
+import com.hanghae99.afterwork.entity.Product;
 import lombok.Data;
 
 import java.util.ArrayList;
