+++ src/main/java/com/hanghae99/afterwork/entity/Category.java
@@ -1,4 +1,4 @@
-package com.hanghae99.afterwork.model;
+package com.hanghae99.afterwork.entity;
 
 import lombok.*;
