+++ src/main/java/com/hanghae99/afterwork/repository/FeedbackRepository.java
@@ -1,6 +1,6 @@
 package com.hanghae99.afterwork.repository;
 
-import com.hanghae99.afterwork.model.Feedback;
+import com.hanghae99.afterwork.entity.Feedback;
 import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.transaction.annotation.Transactional;
