+++ src/main/java/com/hanghae99/afterwork/repository/InterestRepository.java
@@ -1,7 +1,7 @@
 package com.hanghae99.afterwork.repository;
 
-import com.hanghae99.afterwork.model.Interest;
-import com.hanghae99.afterwork.model.User;
+import com.hanghae99.afterwork.entity.Interest;
+import com.hanghae99.afterwork.entity.User;
 import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.transaction.annotation.Transactional;
