+++ src/main/java/com/hanghae99/afterwork/repository/UserRepository.java
@@ -1,6 +1,6 @@
 package com.hanghae99.afterwork.repository;
 
-import com.hanghae99.afterwork.model.User;
+import com.hanghae99.afterwork.entity.User;
 import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.transaction.annotation.Transactional;
