+++ src/test/java/com/hanghae99/afterwork/controller/CollectControllerTest.java
@@ -4,8 +4,8 @@ import com.fasterxml.jackson.core.type.TypeReference;
 import com.fasterxml.jackson.databind.ObjectMapper;
 import com.hanghae99.afterwork.dto.CollectRequestDto;
 import com.hanghae99.afterwork.dto.ProductResponseDto;
+import com.hanghae99.afterwork.entity.*;
 import com.hanghae99.afterwork.exception.ResourceNotFoundException;
-import com.hanghae99.afterwork.model.*;
 import com.hanghae99.afterwork.repository.CategoryRepository;
 import com.hanghae99.afterwork.repository.CollectRepository;
 import com.hanghae99.afterwork.repository.ProductRepository;
