+++ src/test/java/com/hanghae99/afterwork/controller/RecommendControllerTest.java
@@ -3,8 +3,8 @@ package com.hanghae99.afterwork.controller;
 import com.fasterxml.jackson.core.type.TypeReference;
 import com.fasterxml.jackson.databind.ObjectMapper;
 import com.hanghae99.afterwork.dto.ProductResponseDto;
+import com.hanghae99.afterwork.entity.*;
 import com.hanghae99.afterwork.exception.ResourceNotFoundException;
-import com.hanghae99.afterwork.model.*;
 import com.hanghae99.afterwork.repository.*;
 import org.junit.jupiter.api.AfterEach;
 import org.junit.jupiter.api.BeforeEach;
