+++ src/test/java/com/hanghae99/afterwork/controller/SearchControllerTest.java
@@ -1,7 +1,7 @@
 package com.hanghae99.afterwork.controller;
 
-import com.hanghae99.afterwork.model.Category;
-import com.hanghae99.afterwork.model.Product;
+import com.hanghae99.afterwork.entity.Category;
+import com.hanghae99.afterwork.entity.Product;
 import com.hanghae99.afterwork.repository.CategoryRepository;
 import com.hanghae99.afterwork.repository.ProductRepository;
 import com.hanghae99.afterwork.service.CategoryService;
