+++ src/test/java/com/hanghae99/afterwork/controller/UserControllerTest.java
@@ -3,9 +3,9 @@ package com.hanghae99.afterwork.controller;
 import com.fasterxml.jackson.databind.ObjectMapper;
 import com.hanghae99.afterwork.dto.UserRequestDto;
 import com.hanghae99.afterwork.dto.UserResponseDto;
+import com.hanghae99.afterwork.entity.AuthProvider;
+import com.hanghae99.afterwork.entity.User;
 import com.hanghae99.afterwork.exception.ResourceNotFoundException;
-import com.hanghae99.afterwork.model.AuthProvider;
-import com.hanghae99.afterwork.model.User;
 import com.hanghae99.afterwork.repository.UserRepository;
 import org.junit.jupiter.api.AfterEach;
 import org.junit.jupiter.api.BeforeEach;
